Explicitly disable all health checks on the session.
Some scripts may cause a short network outage and we don't want a short
failing serfHealth to invalidate our lock for that.  We rely on the TTL
to invalidate the session in case of disasters.

This was causing rebootmgr to exit when systemd-networkd is restarted
when updating systemd, or when the Mellanox firmware is updated.

Update version to 0.0.28.